+++ packages/dashboard-core-plugins/src/WidgetLoaderPlugin.tsx
@@ -15,10 +15,10 @@ import Log from '@deephaven/log';
 import {
   isWidgetPlugin,
   isWidgetMiddlewarePlugin,
+  createChainedComponent,
   usePlugins,
   type WidgetPlugin,
   type WidgetMiddlewarePlugin,
-  type WidgetComponentProps,
   type WidgetPanelProps,
   type WidgetMiddlewarePanelProps,
 } from '@deephaven/plugin';
@@ -35,43 +35,6 @@ interface WidgetTypeInfo {
   basePlugin: WidgetPlugin;
   /** Middleware plugins to apply, in order from outermost to innermost */
   middleware: WidgetMiddlewarePlugin[];
-}
-
-/**
- * Creates a component that chains middleware around a base component.
- * Each middleware wraps the next, with the base component at the innermost layer.
- */
-function createChainedComponent<T>(
-  baseComponent: React.ComponentType<WidgetComponentProps<T>>,
-  middleware: WidgetMiddlewarePlugin<T>[]
-): React.ComponentType<WidgetComponentProps<T>> {
-  if (middleware.length === 0) {
-    return baseComponent;
-  }
-
-  // Build the chain from inside out (base component is innermost)
-  // Middleware is ordered outermost to innermost, so we reverse to build from inside out
-  return [...middleware]
-    .reverse()
-    .reduce<React.ComponentType<WidgetComponentProps<T>>>(
-      (WrappedComponent, middlewarePlugin) => {
-        const MiddlewareComponent = middlewarePlugin.component;
-
-        function ChainedComponent(props: WidgetComponentProps<T>) {
-          return (
-            // eslint-disable-next-line react/jsx-props-no-spreading
-            <MiddlewareComponent {...props} Component={WrappedComponent} />
-          );
-        }
-        ChainedComponent.displayName = `${middlewarePlugin.name}(${
-          (WrappedComponent as React.ComponentType).displayName ??
-          (WrappedComponent as React.ComponentType).name ??
-          'Component'
-        })`;
-        return ChainedComponent;
-      },
-      baseComponent
-    );
 }
 
 /**
