+++ packages/plugin/src/PluginUtils.tsx
@@ -1,4 +1,4 @@
-import { isValidElement } from 'react';
+import React, { isValidElement } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { getThemeKey, type ThemeData } from '@deephaven/components';
 import { vsPreview } from '@deephaven/icons';
@@ -12,6 +12,8 @@ import {
   isElementPlugin,
   type ElementPlugin,
   type ElementMap,
+  type WidgetMiddlewarePlugin,
+  type WidgetComponentProps,
 } from './PluginTypes';
 
 const log = Log.module('@deephaven/plugin.PluginUtils');
@@ -96,4 +98,41 @@ export function getPluginsElementMap(pluginMap: PluginModuleMap): ElementMap {
   return new Map(
     elementPluginEntries.flatMap(([, plugin]) => Object.entries(plugin.mapping))
   );
+}
+
+/**
+ * Creates a component that chains middleware around a base component.
+ * Each middleware wraps the next, with the base component at the innermost layer.
+ */
+export function createChainedComponent<T>(
+  baseComponent: React.ComponentType<WidgetComponentProps<T>>,
+  middleware: WidgetMiddlewarePlugin<T>[]
+): React.ComponentType<WidgetComponentProps<T>> {
+  if (middleware.length === 0) {
+    return baseComponent;
+  }
+
+  // Build the chain from inside out (base component is innermost)
+  // Middleware is ordered outermost to innermost, so we reverse to build from inside out
+  return [...middleware]
+    .reverse()
+    .reduce<React.ComponentType<WidgetComponentProps<T>>>(
+      (WrappedComponent, middlewarePlugin) => {
+        const MiddlewareComponent = middlewarePlugin.component;
+
+        function ChainedComponent(props: WidgetComponentProps<T>) {
+          return (
+            // eslint-disable-next-line react/jsx-props-no-spreading
+            <MiddlewareComponent {...props} Component={WrappedComponent} />
+          );
+        }
+        ChainedComponent.displayName = `${middlewarePlugin.name}(${
+          (WrappedComponent as React.ComponentType).displayName ??
+          (WrappedComponent as React.ComponentType).name ??
+          'Component'
+        })`;
+        return ChainedComponent;
+      },
+      baseComponent
+    );
 }
