+++ packages/plugin/src/WidgetView.tsx
@@ -1,12 +1,12 @@
-import React, { useMemo } from 'react';
+import { useMemo } from 'react';
 import usePlugins from './usePlugins';
 import {
   isWidgetPlugin,
   isWidgetMiddlewarePlugin,
   type WidgetPlugin,
   type WidgetMiddlewarePlugin,
-  type WidgetComponentProps,
 } from './PluginTypes';
+import { createChainedComponent } from './PluginUtils';
 
 export type WidgetViewProps = {
   /** Fetch function to return the widget */
@@ -15,43 +15,6 @@ export type WidgetViewProps = {
   /** Type of the widget */
   type: string;
 };
-
-/**
- * Creates a component that chains middleware around a base component.
- * Each middleware wraps the next, with the base component at the innermost layer.
- */
-function createChainedComponent<T>(
-  baseComponent: React.ComponentType<WidgetComponentProps<T>>,
-  middleware: WidgetMiddlewarePlugin<T>[]
-): React.ComponentType<WidgetComponentProps<T>> {
-  if (middleware.length === 0) {
-    return baseComponent;
-  }
-
-  // Build the chain from inside out (base component is innermost)
-  // Middleware is ordered outermost to innermost, so we reverse to build from inside out
-  return [...middleware]
-    .reverse()
-    .reduce<React.ComponentType<WidgetComponentProps<T>>>(
-      (WrappedComponent, middlewarePlugin) => {
-        const MiddlewareComponent = middlewarePlugin.component;
-
-        function ChainedComponent(props: WidgetComponentProps<T>) {
-          return (
-            // eslint-disable-next-line react/jsx-props-no-spreading
-            <MiddlewareComponent {...props} Component={WrappedComponent} />
-          );
-        }
-        ChainedComponent.displayName = `${middlewarePlugin.name}(${
-          (WrappedComponent as React.ComponentType).displayName ??
-          (WrappedComponent as React.ComponentType).name ??
-          'Component'
-        })`;
-        return ChainedComponent;
-      },
-      baseComponent
-    );
-}
 
 export function WidgetView({ fetch, type }: WidgetViewProps): JSX.Element {
   const plugins = usePlugins();
